Revert "Htsang/update bot (#155)"
This reverts commit a4d3d1d3aa9b6310e94e6ccef6e3f6b01f107fc7.
