+++ server/bot-service.js
@@ -7,14 +7,6 @@ const taskService = require("./task-service");
 const userService = require("./user-service");
 const groupService = require("./group-service");
 const utils = require("./utils");
-const {
-  createTaskBotPrompt,
-  signOutBotResponse,
-  SSOResponse,
-  defaultAdaptiveCardAttachment,
-  getAdaptiveCardForTask,
-  getTaskCardWithPreview,
-} = require("./bot-utils");
 
 class AuthBot extends builder.UniversalBot {
   constructor(_connector, botSettings) {
@@ -127,9 +119,9 @@ class AuthBot extends builder.UniversalBot {
       } else if (event.name === "composeExtension/query") {
         await this.handleQuery(cb, event, session);
       } else if (event.name === "composeExtension/fetchTask") {
-        await this.handleFetchTask(cb, event, session);
+        await this.handleFetchTask(cb, session);
       } else if (event.name === "composeExtension/submitAction") {
-        await this.handleSubmitAction(cb, event);
+        //do nothing when the compose extension close button is pressed
       } else {
         // Simulate a normal message and route it, but remember the original invoke message
         const payload = event.value;
@@ -149,91 +141,44 @@ class AuthBot extends builder.UniversalBot {
     cb(null, "");
   }
 
-  async handleSubmitAction(cb, event) {
-    const { value, address, sourceEvent } = event;
-    if (value?.commandId === "createTask" && value?.data?.title && address) {
-      const result = await this.handleCreateTask(
-        value.data,
-        address,
-        sourceEvent
-      );
-      cb(null, result);
-    } else {
-      cb(null);
-    }
-  }
-
-  async handleCreateTask(data, address, sourceEvent) {
-    const response = {
-      composeExtension: {
-        attachmentLayout: "list",
-        type: "result",
-        attachments: [],
+  // Handle fetch task
+  async handleFetchTask(cb, session) {
+    utils.setUserToken(session, null);
+    const card = {
+      contentType: "application/vnd.microsoft.card.adaptive",
+      content: {
+        version: "1.0.0",
+        type: "AdaptiveCard",
+        body: [
+          {
+            type: "TextBlock",
+            text: "You have been signed out.",
+          },
+        ],
+        actions: [
+          {
+            type: "Action.Submit",
+            title: "Close",
+            data: {
+              key: "close",
+            },
+          },
+        ],
       },
     };
 
-    const taskTitle = data.title;
-    const userId = address?.user?.aadObjectId;
-    const { conversationType } = address.conversation;
-    // Since tasks are organized at the team level (no channel tasks), use team id for channel posts,
-    // otherwise, use conversation ID.
-    const threadId =
-      conversationType == "channel"
-        ? sourceEvent?.team?.id
-        : address?.conversation?.id;
-
-    try {
-      if (conversationType !== "personal") {
-        let group = await groupService.get(threadId);
-        if (!group) {
-          group = await groupService.create(threadId, address.serviceUrl);
-        }
-        const members = await getMembers(group.serviceUrl, threadId);
-        if (members && members.some((member) => member.objectId === userId)) {
-          const task = await taskService.createForGroup(group.id, taskTitle);
-          const card = getAdaptiveCardForTask(task, undefined, true);
-          response.composeExtension.attachments = [card];
-          return response;
-        } else {
-          console.error(
-            `There are no group members or the user is not one of them.`
-          );
-          response.composeExtension.attachments = [
-            defaultAdaptiveCardAttachment,
-          ];
-          return response;
-        }
-      } else {
-        const task = await taskService.createForUser(userId, taskTitle);
-        const card = getAdaptiveCardForTask(task);
-        response.composeExtension.attachments = [card];
-        return response;
-      }
-    } catch (error) {
-      console.error(error);
-      response.composeExtension.attachments = [defaultAdaptiveCardAttachment];
-      return response;
-    }
-  }
-
-  // Handle fetch task
-  async handleFetchTask(cb, event, session) {
-    if (event.value.commandId === "SignOutCommand") {
-      this.handleSignOutCommand(cb, session);
-    } else if (event.value.commandId === "createTask") {
-      this.handleCreateTaskCommand(cb);
-    } else {
-      cb(null);
-    }
-  }
-
-  handleCreateTaskCommand(cb) {
-    cb(null, createTaskBotPrompt);
-  }
-
-  handleSignOutCommand(cb, session) {
-    utils.setUserToken(session, null);
-    cb(null, signOutBotResponse);
+    const response = {
+      task: {
+        type: "continue",
+        value: {
+          card: card,
+          heigth: 200,
+          width: 400,
+          title: "Adaptive Card: Inputs",
+        },
+      },
+    };
+    cb(null, response);
   }
 
   // Handle query
@@ -245,25 +190,28 @@ class AuthBot extends builder.UniversalBot {
 
     //check if user is authenticated if not send the sign in request
     if (!utils.getUserToken(session)) {
-      cb(null, SSOResponse);
-      return;
+      cb(null, this.getSSOResponse());
     }
 
+    const user = await userService.getUser(
+      session.message.address.user.aadObjectId
+    );
     const searchString =
       event.value.parameters.length > 0 && event.value.parameters[0].value;
-
-    const tasks = await this.getTasksInContext(session, event);
-
-    const groupPath =
-      event.address.conversationType == "personal" ? "" : "/group";
-
+    const tasks = await taskService.getForUser(user._id);
     const attachments = tasks
-      .filter(
-        (task) =>
-          task.title &&
-          task.title.toLowerCase().indexOf(searchString.toLowerCase()) >= 0
-      )
-      .map((task) => getTaskCardWithPreview(task, groupPath));
+      .filter((task) => task.title.indexOf(searchString) >= 0)
+      .map((task) => ({
+        content: {
+          title: task.title,
+          images: [
+            {
+              url: `${process.env.APPSETTING_AAD_BaseUri}/checkmark.png`,
+            },
+          ],
+        },
+        contentType: "application/vnd.microsoft.card.thumbnail",
+      }));
 
     const result = {
       attachmentLayout: "list",
@@ -278,26 +226,6 @@ class AuthBot extends builder.UniversalBot {
     cb(null, response);
   }
 
-  async getTasksInContext(session, event) {
-    const { address, sourceEvent } = event;
-    if (address.conversationType !== "personal") {
-      // Since tasks are organized at the team level (no channel tasks), use team id for channel posts,
-      // otherwise, use conversation ID.
-      const threadId =
-        address.conversationType == "channel"
-          ? sourceEvent?.team?.id
-          : address?.conversation?.id;
-
-      return groupService
-        .get(threadId)
-        .then((group) => taskService.getForGroup(group._id));
-    } else {
-      return userService
-        .getUser(session.message.address.user.aadObjectId)
-        .then((user) => taskService.getForUser(user.id));
-    }
-  }
-
   // Handle Link Unfurl
   async handleLinkUnfurl(cb, event, session) {
     //store access token
@@ -308,32 +236,36 @@ class AuthBot extends builder.UniversalBot {
     //check if user is authenticated if not send the sign in request
     if (!utils.getUserToken(session)) {
       cb(null, this.getSSOResponse());
-      return;
     }
 
     const urlObj = new URL(event.value.url);
     const taskId = urlObj.searchParams.get("task");
-    const shareTag = urlObj.searchParams.get("shareTag");
     if (taskId) {
       const taskObj = await taskService.get(taskId);
-      const attachment = getAdaptiveCardForTask(taskObj, shareTag, true);
-      const result = {
-        attachmentLayout: "list",
-        type: "result",
-        attachments: [attachment],
-        responseType: "composeExtension",
-      };
-      const response = {
-        composeExtension: result,
-      };
-      cb(null, response);
-      return;
+      if (taskObj) {
+        const attachment = this.getAdaptiveCardForTask(
+          event.value.url,
+          taskObj
+        );
+        const result = {
+          attachmentLayout: "list",
+          type: "result",
+          attachments: [attachment],
+          responseType: "composeExtension",
+        };
+        const response = {
+          composeExtension: result,
+        };
+        cb(null, response);
+        return;
+      }
     }
 
+    const attachment = this.getAdaptiveCardAttachment();
     const result = {
       attachmentLayout: "list",
       type: "result",
-      attachments: [defaultAdaptiveCardAttachment],
+      attachments: [attachment],
       responseType: "composeExtension",
     };
     const response = {
@@ -348,6 +280,166 @@ class AuthBot extends builder.UniversalBot {
       accessToken: authentication.token,
     };
     utils.setUserToken(session, token);
+  }
+
+  getSSOResponse() {
+    return {
+      composeExtension: {
+        type: "silentAuth",
+        responseType: "composeExtension",
+        suggestedActions: {},
+      },
+    };
+  }
+  getAdaptiveCardAttachment() {
+    const contentUrl = "https://taskmeow.com/group/?inTeamsSSO=true";
+    const websiteUrl = "https://taskmeow.com/group";
+
+    const adaptiveCardJson = {
+      contentType: "application/vnd.microsoft.card.adaptive",
+      content: {
+        type: "AdaptiveCard",
+        version: "1.0",
+        body: [
+          {
+            type: "TextBlock",
+            size: "Medium",
+            weight: "Bolder",
+            text: "Publish Adaptive Card Schema",
+          },
+          {
+            type: "ColumnSet",
+            columns: [
+              {
+                type: "Column",
+                items: [
+                  {
+                    type: "TextBlock",
+                    weight: "Bolder",
+                    text: "Name",
+                    wrap: true,
+                  },
+                  {
+                    type: "TextBlock",
+                    spacing: "None",
+                    text: "Created today",
+                    isSubtle: true,
+                    wrap: true,
+                  },
+                ],
+                width: "stretch",
+              },
+            ],
+          },
+          {
+            type: "TextBlock",
+            text: "Description",
+            wrap: true,
+          },
+        ],
+        actions: [
+          {
+            type: "Action.OpenUrl",
+            title: "Outside Teams",
+            url: contentUrl,
+          },
+          {
+            type: "Action.Submit",
+            title: "View",
+            data: {
+              msteams: {
+                type: "invoke",
+                value: {
+                  type: "tab/tabInfoAction",
+                  tabInfo: {
+                    contentUrl: contentUrl,
+                    websiteUrl: websiteUrl,
+                    name: "Tasks",
+                    entityId: "entityId",
+                  },
+                },
+              },
+            },
+          },
+        ],
+      },
+      preview: {
+        content: {
+          title: "Description",
+          text: "Task",
+          images: [
+            {
+              url: "https://taskmeow.com/static/media/logo.28c3e78f.svg",
+            },
+          ],
+        },
+        contentType: "application/vnd.microsoft.card.thumbnail",
+      },
+    };
+    return adaptiveCardJson;
+  }
+
+  getAdaptiveCardForTask(url, task) {
+    const contentUrl = `${url}&inTeamsSSO=true`;
+    const websiteUrl = url;
+    const adaptiveCardJson = {
+      contentType: "application/vnd.microsoft.card.adaptive",
+      content: {
+        type: "AdaptiveCard",
+        version: "1.0",
+        body: [
+          {
+            type: "TextBlock",
+            size: "Medium",
+            weight: "Bolder",
+            text: task.title,
+          },
+          {
+            type: "TextBlock",
+            text: "Description",
+            wrap: true,
+          },
+        ],
+        actions: [
+          {
+            type: "Action.OpenUrl",
+            title: "Outside Teams",
+            url: websiteUrl,
+          },
+          {
+            type: "Action.Submit",
+            title: "View",
+            data: {
+              msteams: {
+                type: "invoke",
+                value: {
+                  type: "tab/tabInfoAction",
+                  tabInfo: {
+                    contentUrl: contentUrl,
+                    websiteUrl: websiteUrl,
+                    name: "Tasks",
+                    entityId: "entityId",
+                  },
+                },
+              },
+            },
+          },
+        ],
+      },
+      preview: {
+        content: {
+          title: "Description",
+          text: "Task",
+          images: [
+            {
+              url: "https://taskmeow.com/static/media/logo.28c3e78f.svg",
+            },
+          ],
+        },
+        contentType: "application/vnd.microsoft.card.thumbnail",
+      },
+    };
+    return adaptiveCardJson;
   }
 }
